Add module 5

+++ src/main/java/com/example/demo/service/EmployeeService.java
@@ -7,28 +7,31 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 
+
 @Service
-@RequiredArgsConstructor
 public class EmployeeService {
     private final EmployeeRepository employeeRepository;
+    public EmployeeService(EmployeeRepository employeeRepository) {
+        this.employeeRepository = employeeRepository;
+    }
 
     public List<Employee> getAllEmployees() {
         return employeeRepository.findAll();
     }
 
-    public Employee addEmployee(Employee employee) {
+    public Employee createEmployee(Employee employee) {
         return employeeRepository.save(employee);
     }
 
+    public Employee getEmployeeById(Long id) {
+        return employeeRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Không tìm thấy nhân viên với ID: " + id));
+    }
+
     public void deleteEmployee(Long id) {
+        if (!employeeRepository.existsById(id)) {
+            throw new RuntimeException("Không tìm thấy nhân viên với ID: " + id);
+        }
         employeeRepository.deleteById(id);
-    }
-
-    public List<Employee> searchByName(String name) {
-        return employeeRepository.findByNameContainingIgnoreCase(name);
-    }
-
-    public List<Employee> searchByDepartment(String departmentName) {
-        return employeeRepository.findByDepartment_NameContainingIgnoreCase(departmentName);
     }
 }
